Multiple Attempts and failures detailed.
Just wanted to check in.
attached click event to the leave band button,
attempted: when leave band btn clicked, wanted to call the disconnect
event from app sever in hopes to broadcast which band member
disconnected to all the sockets, so the band members could be removed
or added to the views. The server socket didn’t seem to be callable
from the client side. i.e. socket.broadcast.emit("disconnect”); was not
hitting the server app.js disconnect event.
attempt 2:
Noticed that the ID of the user is emitted from server to all client
sockets, attempted to utilize the connect and disconnect broadcast to
pass more information to the clients, band information and userName
etc, then jquery  append the html in the connect socket on the client.
Realized that express or embedded JS should be researched to display
the view dynamically instead of the jQuery append.
Checked in before exploring that route.